+++ src/main.cpp
@@ -1,7 +1,7 @@
 #include"header.hpp"
 #include"ui.hpp"
 CUI ui;
-#include"private.hpp"
+#include"pvtModules.hpp"
 auto main(i32 argc,cstr argv[])->i32{
     bool optError{};
     if(argc>1){
